+++ Identity.Base.Organizations/Extensions/IdentityBaseBuilderOrganizationsExtensions.cs
@@ -63,4 +63,22 @@ public static class IdentityBaseBuilderOrganizationsExtensions
         builder.Services.AddScoped<IOrganizationCreationListener, TListener>();
         return builder;
     }
+
+    public static IdentityBaseBuilder AddOrganizationUpdateListener<TListener>(this IdentityBaseBuilder builder)
+        where TListener : class, IOrganizationUpdateListener
+    {
+        ArgumentNullException.ThrowIfNull(builder);
+
+        builder.Services.AddScoped<IOrganizationUpdateListener, TListener>();
+        return builder;
+    }
+
+    public static IdentityBaseBuilder AddOrganizationArchiveListener<TListener>(this IdentityBaseBuilder builder)
+        where TListener : class, IOrganizationArchiveListener
+    {
+        ArgumentNullException.ThrowIfNull(builder);
+
+        builder.Services.AddScoped<IOrganizationArchiveListener, TListener>();
+        return builder;
+    }
 }